Fix: More resilient native type mapping
Swapped out the hardcoded "minecraft:grass" fallback for a `DEFAULT_FALLBACK` constant and use it instead. This should prevent issues if "minecraft:grass" ever needs to be mapped itself. Also, use this `DEFAULT_FALLBACK` constant to replace the `minecraft:grass` if exist. Because, `minecraft:grass` dosen't exist in the minecraft namespaces...or registry thingy